+++ arrow/src/main/java9/module-info.java
@@ -1,0 +1,6 @@
+module org.bytedeco.arrow {
+  requires transitive org.bytedeco.javacpp;
+  exports org.bytedeco.arrow.global;
+  exports org.bytedeco.arrow.presets;
+  exports org.bytedeco.arrow;
+}